backend/pyproject.toml
[tool.poetry]
name = "todo-backend"
version = "0.1.0"
description = "Secure Todo Backend API with JWT authentication"
authors = ["Developer <developer@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.104.1"
sqlmodel = "^0.0.16"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
pyjwt = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
psycopg2-binary = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
cryptography = "^41.0.8"
httpx = "^0.25.2"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.flake8]
max-line-length = 88
extend-ignore = ['E203', 'W503']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"